PICK.1: picker transient-mode extension
Adds the transient-mode data types and interaction layer for
grouped action menus within the picker. Magit is the first
consumer (dispatch menus, branch menus, stash menus), but the
mechanism is general-purpose.

Data types (lattice-picker::transient):
- TransientSpec, TransientGroup, TransientItem, TransientItemKind
- TransientState, TransientValue
- transient_initial_state(spec) -> TransientState

Picker struct extension:
- transient: Option<Arc<TransientSpec>>
- transient_state: TransientState
- transient_stack: Vec<(Arc<TransientSpec>, TransientState)>

Editor API (lattice-host):
- open_transient(spec) - seats a picker in transient mode
- do_transient_trigger(key) - matches key against items, dispatches
- do_transient_toggle_flag(name) - toggles boolean flag
- close_transient() - dismisses via do_picker_dismiss

Input routing (dispatch.rs):
- PickerAppend checks transient.is_some() -> routes to transient trigger
- PickerBackspace in transient mode -> pops stack (back to parent)
- New Action variants: TransientTrigger(String), TransientToggleFlag(String)

Deferred to MG.8:
- Action handler invocation (needs ActionContext from mode Guard)
- Argument minibuffer prompt flow
- Per-item rendering (TUI + GPUI grouped section layout)

62 picker tests + 714 host tests pass green.

+++ crates/lattice-picker/src/picker_sources.rs
@@ -769,7 +769,7 @@ impl PickerSourceGenerator for LinesSource {
             // Slice 7b.4: typed accept payload.
             cand.accept_action = Some(Box::new(lattice_completion::AcceptAction::JumpInBuffer {
                 buffer_id: lattice_core::BufferId(buffer_id),
-                line: line as u32,
+                line: line,
                 col: 0,
             }));
             pairs.push((
